# docs/swagger.yaml
basePath: /
definitions:
...
  input.AddressDeleteAddressReq:
    properties:
      id:
        example: 1
        type: integer
    required:
    - id
    type: object
  input.AddressGetAddressByIDReq:
    properties:
      id:
        example: 1
        type: integer
    required:
    - id
    type: object
  input.AddressGetAddressListReq:
    properties:
      type:
        enum:
        - 0
        - 1
        example: 1
        type: integer
    type: object
  input.AddressGetDefaultAddressReq:
    properties:
      type:
        enum:
        - 0
        - 1
        example: 1
        type: integer
    type: object
  input.AddressSaveAddressReq:
    properties:
      detail:
        example: xxxxxxxxxx
        type: string
      email:
        example: xxxxx@xxx.xx
        type: string
      id:
        example: 2
        type: integer
      is_default:
        enum:
        - 0
        - 1
        example: 1
        type: integer
      name:
        example: 哎呀又胖啦
        type: string
      phone:
        example: "18888888888"
        type: string
      region:
        example: xxx xxx xxx
        type: string
      region_code:
        example: '[''370000'', ''370100'', ''370116'']'
        type: string
      type:
        enum:
        - 0
        - 1
        example: 1
        type: integer
    type: object
...
  resp.AddressGetAddressByIDResp:
    properties:
      detail:
        description: 详细地址
        example: xxxxxxxxxxxxxxx
        type: string
      email:
        description: 邮箱地址
        example: xxxxxxxx@xxx.xx
        type: string
      id:
        description: id
        example: 1
        type: integer
      is_default:
        description: 默认地址
        enum:
        - 0
        - 1
        example: 1
        type: integer
      name:
        description: 收货人姓名
        example: 哎呀又胖啦
        type: string
      phone:
        description: 收货人手机号
        example: "18888888888"
        type: string
      region:
        description: 地区文字描述
        example: xxxxx xxxxx xxxxx
        type: string
      region_code:
        description: 地区code
        example: '[''370000'', ''370100'', ''370116'']'
        type: string
      type:
        description: 类型
        enum:
        - 0
        - 1
        example: 1
        type: integer
    type: object
  resp.AddressGetAddressListResp:
    properties:
      list:
        description: 收货地址列表
        items:
          $ref: '#/definitions/resp.AddressItem'
        type: array
    type: object
  resp.AddressGetDefaultAddressResp:
    properties:
      detail:
        description: 详细地址
        example: xxxxxxxxxxxxxxx
        type: string
      email:
        description: 邮箱地址
        example: xxxxxxxx@xxx.xx
        type: string
      id:
        description: id
        example: 1
        type: integer
      is_default:
        description: 默认地址
        enum:
        - 0
        - 1
        example: 1
        type: integer
      name:
        description: 收货人姓名
        example: 哎呀又胖啦
        type: string
      phone:
        description: 收货人手机号
        example: "18888888888"
        type: string
      region:
        description: 地区文字描述
        example: xxxxx xxxxx xxxxx
        type: string
      region_code:
        description: 地区code
        example: '[''370000'', ''370100'', ''370116'']'
        type: string
      type:
        description: 类型
        enum:
        - 0
        - 1
        example: 1
        type: integer
    type: object
  resp.AddressItem:
    properties:
      detail:
        description: 详细地址
        example: xxxxxxxxxxxxxxx
        type: string
      email:
        description: 邮箱地址
        example: xxxxxxxx@xxx.xx
        type: string
      id:
        description: id
        example: 1
        type: integer
      is_default:
        description: 默认地址
        enum:
        - 0
        - 1
        example: 1
        type: integer
      name:
        description: 收货人姓名
        example: 哎呀又胖啦
        type: string
      phone:
        description: 收货人手机号
        example: "18888888888"
        type: string
      region:
        description: 地区文字描述
        example: xxxxx xxxxx xxxxx
        type: string
      region_code:
        description: 地区code
        example: '[''370000'', ''370100'', ''370116'']'
        type: string
      type:
        description: 类型
        enum:
        - 0
        - 1
        example: 1
        type: integer
    type: object
...
  response.Response:
    properties:
      code:
        description: 响应code
        example: 0
        type: integer
      data:
        description: 响应数据
      msg:
        description: 响应信息
        example: success
        type: string
    type: object
info:
  contact:
    email: junjie.he.925@gmail.com
    name: 何俊杰
  description: BiliLive Assistant 系统接口文档
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: BiliLive Assistant API
  version: "1.0"
paths:
...
  /api/shop/address/default:
    post:
      description: 返回当前用户在指定收货地址类型（0 虚拟，1 实体）下的默认收货地址，供下单确认等页面默认回填使用；该类型下暂无默认地址时返回空对象。
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/input.AddressGetDefaultAddressReq'
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.AddressGetDefaultAddressResp'
              type: object
      security:
      - BearerAuth: []
      summary: 获取指定收货地址类型的默认地址
      tags:
      - 移动端
  /api/shop/address/delete:
    post:
      description: 删除当前用户的某条收货地址（软删除）；仅可删除归属当前用户本人且存在的地址，否则返回错误。
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/input.AddressDeleteAddressReq'
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除收货地址
      tags:
      - 移动端
  /api/shop/address/detail:
    post:
      description: 返回当前用户某条收货地址的完整信息，常用于编辑前的信息回显；仅可查询归属当前用户本人的地址，否则返回错误。
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/input.AddressGetAddressByIDReq'
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.AddressGetAddressByIDResp'
              type: object
      security:
      - BearerAuth: []
      summary: 获取收货地址详情
      tags:
      - 移动端
  /api/shop/address/list:
    post:
      description: 按收货地址类型获取当前用户的全部收货地址，不分页返回，常用于地址管理页面的列表展示；type 缺省时返回全部类型。
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      - description: 请求参数（可省略；省略或未传 type 时返回全部类型）
        in: body
        name: data
        schema:
          $ref: '#/definitions/input.AddressGetAddressListReq'
      responses:
        "200":
          description: 统一响应（code=0成功，其它失败）
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/resp.AddressGetAddressListResp'
              type: object
      security:
      - BearerAuth: []
      summary: 获取收货地址列表
      tags:
      - 移动端
  /api/shop/address/save:
    post:
      description: 新增或修改当前用户的收货地址：携带 id 视为修改该条地址，不携带 id 视为新增；修改时省略的字段保留原值，is_default
        为 1 时该地址被设为对应类型下的默认地址。
      parameters:
      - default: zh
        description: '语言标识（zh: 中文，en: English）'
        enum:
        - zh
        - en
        in: header
        name: Accept-Language
        type: string
      - description: 请求参数
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/input.AddressSaveAddressReq'
      responses:
        "200":
          description: 统一响应（code=0成功，data 为地址 ID）
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  format: int64
                  type: integer
              type: object
      security:
      - BearerAuth: []
      summary: 新增/修改收货地址
      tags:
      - 移动端
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"